+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 # Changelog
+
+#### x.x.x
+  * added translations to goodmap
+
 #### 0.4.0
   * changed API: /api/data pin attribute "data" is served as an array of pairs instead of dict